Finished both Pages of Treatment, some more fix ups with Patients.
Appointments is still left to code. Then some tweaks to the Wire frames need to be done. Then i need to format it in a document and write some documentations and annotations and I'm all done! I haven't even had the assesment for a week yet and the hardest part is almost over.